+++ src/pages/EditList.jsx
@@ -1,14 +1,10 @@
+import { useState, useEffect } from "react";
+import { useTheme } from "@mui/material";
 import React from "react";
-import { useState, useEffect, useContext } from "react";
 import axios from "axios";
 import Card from "../components/Card";
 import Box from "@mui/material/Box";
-import Textfield from "../components/Textfield";
-import Pagination from "@mui/material/Pagination";
-import { Link } from "react-router-dom";
 import styled from "styled-components";
-import { ThemeContext } from "../context/ThemeContext";
-import { useTheme } from "@mui/material";
 
 const StyledBox = styled.div`
   height: 100%;
@@ -17,32 +13,26 @@ const StyledBox = styled.div`
   align-items: center;
 `;
 
-export default function Edycja() {
+export default function EditList() {
   const [url, setUrl] = useState(
     `https://pokeapi.co/api/v2/pokemon/?limit=150&offset=0`
   );
   const [pokedex, setPokedex] = useState([]);
-  const [search, setSearch] = useState();
   const [error, setError] = useState(null);
   const theme = useTheme();
-  const colorMode = useContext(ThemeContext);
 
   useEffect(() => {
     const fetchData = async () => {
       try {
         const response = await axios.get(url);
-        const data = response.data.results;
-
-        search
-          ? setPokedex(searchFilter(data, search))
-          : setPokedex(response.data.results);
+        setPokedex(response.data.results);
       } catch (error) {
         setError(error);
       }
     };
 
     fetchData();
-  }, [url, search]);
+  }, [url]);
 
   return (
     <StyledBox style={{ backgroundColor: theme.palette.background.contrast }}>
