+++ src/pages/Register.jsx
@@ -1,91 +1,46 @@
-import { useState, useEffect, useContext } from "react";
+import { useState } from "react";
 import { useSnackbar } from "notistack";
+import { Formik, Form, Field, ErrorMessage } from "formik";
+import { Box, Button, TextField } from "@mui/material";
+import { useTheme } from "@mui/material";
 import axios from "axios";
 import * as Yup from "yup";
-import { Formik, Form, Field, ErrorMessage } from "formik";
-import { Box, Button, Container, TextField, Typography } from "@mui/material";
-import { useNavigate } from "react-router-dom";
-import { useTheme } from "@mui/material";
-import styled from "styled-components";
-
-const StyledBox = styled(Box)`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border: 3px solid black;
-  margin-top: 30px;
-  background-color: ${({ theme }) => theme.palette.background.default};
-  width: 300px;
-  height: 350px;
-
-  & h1 {
-    margin-bottom: 16px;
-  }
-`;
-
-const StyledTextField = styled(TextField)`
-  background-color: ${({ theme }) => theme.palette.background.contrast};
-`;
-
-const StyledButton = styled(Button)`
-  margin-top: 8px;
-`;
 
 const userSchema = Yup.object().shape({
+  name: Yup.string().required("Imię jest wymagane."),
   email: Yup.string()
     .email("Nieprawidłowy adres e-mail.")
     .required("Adres e-mail jest wymagany."),
-  password: Yup.string().required("Hasło jest wymagane."),
+  password: Yup.string()
+    .matches(
+      /^(?=.*[A-Z])(?=.*\d)(?=.*[!@#$%^&*()_+])(?=.*[a-z]).{8,}$/,
+      "Hasło musi składać się z co najmniej 8 znaków, w tym co najmniej 1 dużej litery, 1 cyfry i 1 znaku specjalnego."
+    )
+    .required("Hasło jest wymagane."),
+  confirmPassword: Yup.string()
+    .oneOf([Yup.ref("password"), null], "Hasła muszą się zgadzać.")
+    .required("Potwierdź hasło."),
 });
 
-export default function Logowanie({ setUserData }) {
+export default function Register() {
   const [isSubmitting, setIsSubmitting] = useState(false);
-  const [userInfo, setUserInfo] = useState([]);
-  const [test, setTest] = useState([]);
   const { enqueueSnackbar } = useSnackbar();
-  const navigate = useNavigate();
   const theme = useTheme();
-
-  useEffect(() => {
-    axios
-      .get("http://localhost:3001/user")
-      .then((response) =>
-        setTest(
-          response.data?.filter((item) => {
-            return item.name && item.password;
-          })
-        )
-      )
-      .catch((error) => console.log(error));
-  }, []);
-
-  const handleClick = (text, type) => {
-    enqueueSnackbar(text, { variant: type });
-  };
 
   const handleOnSubmit = (values, actions) => {
     setIsSubmitting(true);
     console.log(values);
     actions.setSubmitting(false);
-    setUserInfo(values);
-    checkLogin(values);
+    axios.post(`http://localhost:3001/user/`, {
+      name: values.name,
+      email: values.email,
+      password: values.password,
+    });
+    handleClick();
   };
 
-  const checkLogin = (userData) => {
-    const currLog = userData.email;
-    const currPass = userData.password;
-    const filtered = test.filter((item) => {
-      return item.email === currLog;
-    });
-
-    if (filtered[0].password === currPass && filtered[0].email === currLog) {
-      localStorage.setItem("userData", JSON.stringify(userData));
-      setUserData(userData);
-      handleClick("Login succes", "success");
-      navigate("/edycja");
-    } else {
-      handleClick("Login faild, wrond password or email", "error");
-    }
+  const handleClick = () => {
+    enqueueSnackbar("Register Succes", { variant: "success" });
   };
 
   return (
@@ -109,15 +64,17 @@ export default function Logowanie({ setUserData }) {
           marginTop: "30px",
         }}
         style={{
-          backgroundColor: theme.palette.background.default,
+          backgroundColor: theme.palette.background.login,
           width: "300px",
-          height: "350px",
+          maxHeight: "550px",
         }}
       >
         <Formik
           initialValues={{
+            name: "",
             email: "",
             password: "",
+            confirmPassword: "",
           }}
           onSubmit={handleOnSubmit}
           validationSchema={userSchema}
@@ -125,7 +82,16 @@ export default function Logowanie({ setUserData }) {
           {({ errors, touched, isSubmitting }) => (
             <Form>
               <Box marginBottom={2}>
-                <h1>Logowanie</h1>
+                <h1>Register</h1>
+              </Box>
+              <Box marginBottom={2}>
+                <Field
+                  as={TextField}
+                  name="name"
+                  label="Imię"
+                  error={touched.name && errors.name ? true : false}
+                  helperText={touched.name && errors.name ? errors.name : ""}
+                />
               </Box>
               <Box marginBottom={2}>
                 <Field
@@ -135,7 +101,6 @@ export default function Logowanie({ setUserData }) {
                   label="Adres e-mail"
                   error={touched.email && errors.email ? true : false}
                   helperText={touched.email && errors.email ? errors.email : ""}
-                  style={{ background: theme.palette.background.contrast }}
                 />
               </Box>
               <Box marginBottom={2}>
@@ -149,7 +114,25 @@ export default function Logowanie({ setUserData }) {
                   helperText={
                     touched.password && errors.password ? errors.password : ""
                   }
-                  style={{ background: theme.palette.background.contrast }}
+                />
+              </Box>
+              <Box marginBottom={2}>
+                <Field
+                  as={TextField}
+                  type="password"
+                  name="confirmPassword"
+                  label="Potwierdź hasło"
+                  fullWidth
+                  error={
+                    touched.confirmPassword && errors.confirmPassword
+                      ? true
+                      : false
+                  }
+                  helperText={
+                    touched.confirmPassword && errors.confirmPassword
+                      ? errors.confirmPassword
+                      : ""
+                  }
                 />
               </Box>
               <Box display="flex" justifyContent="center" marginBottom={2}>
@@ -159,7 +142,7 @@ export default function Logowanie({ setUserData }) {
                   type="submit"
                   disabled={isSubmitting}
                 >
-                  Zaloguj
+                  Register
                 </Button>
               </Box>
             </Form>
