--- src/pages/Wyloguj.jsx
@@ -1,3 +1,0 @@
-export default function Home() {
-  return <h1>wyloguj</h1>;
-}
